refactor(app): split SendTelemetryPanel and IntegrationsDrawer into modules
Break the two large GettingStarted components into directories of focused
files (panel/card/tab, drawer/guide/tile) to keep files under the size limit.
