+++ packages/app/src/components/GettingStarted/IntegrationsDrawer/index.ts
@@ -1,0 +1,4 @@
+export {
+  IntegrationsDrawer,
+  type IntegrationsDrawerProps,
+} from './IntegrationsDrawer';
