+++ packages/app/src/components/GettingStarted/SendTelemetryPanel/index.ts
@@ -1,0 +1,4 @@
+export {
+  SendTelemetryPanel,
+  type SendTelemetryPanelProps,
+} from './SendTelemetryPanel';
